undef $writer to avoid a memory leak
If we don't do this we leak CGI::PSGI instance per
request and other object referenced by it, but
only if plack's response_cb util is used.

I couldn't unwrap everything to find particular
reason of the leak, but this change helps. Test
proves that module leaks without the change and
doesn't leak with.